backend/requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
certifi==2024.2.2
cffi==1.16.0
# channels==3.0.4
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click==8.1.7
constantly==23.10.4
cryptography==42.0.7
daphne==4.1.2
Django==5.0.3
django-cors-headers==4.3.1
django-distill==3.1.3
django-environ==0.11.2
django-redis==5.4.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
msgpack==1.0.8
pillow==10.3.0
priority==1.3.0
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.1.0
redis==5.0.4
requests==2.31.0
service-identity==24.1.0
six==1.16.0
sqlparse==0.4.4
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.29.0
websocket-client==1.8.0
websockets==12.0
zope.interface==6.3
mailjet_rest==1.3.4